Added Gaia DR1, IGSL and ALLWISE astrometric reference catalogs. Added new configuration parameters that had been missing in the output XML-VOTable. Replaced the aserver.cgi "big catalog" interface with VizieR queries. Removed support for obsolete catalogs (all GSC releases before 2.3, USNO-A1, all SDSS releases before R9, all UCAC releases before 4, CMC releases before 15), and fixed some issues with the remaining ones. Server connection timeout reduced to 10s. Added HEADER_NAME and AHEADER_NAME configuration parameters for specifying explicit header names (thanks to R.Gruendl for the suggestion). Added support for @ lists in configuration file and command line. Re-organized the default configuration file. Fixed numerical error in ecliptic coordinates (thanks to H.Bouy for reporting this issue). Pushed version number to 2.4.3.

+++ src/url.h
@@ -7,7 +7,7 @@
 *
 *	This file part of:	SCAMP
 *
-*	Copyright:		(C) 2002-2014 Emmanuel Bertin -- IAP/CNRS/UPMC
+*	Copyright:		(C) 2002-2016 Emmanuel Bertin -- IAP/CNRS/UPMC
 *
 *	License:		GNU General Public License
 *
@@ -24,7 +24,7 @@
 *	This file incorporates work covered by the copyright and permission
 *	notice below.
 *
-*	Last modified:		03/06/2015
+*	Last modified:		31/03/2016
 *
 *%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*/
 /*****************************************************************************
@@ -73,7 +73,7 @@ typedef struct fcurl_data {
 }	URL_FILE;
 
 /*------------------------------- functions ---------------------------------*/
-URL_FILE	*url_fopen(const char *url, long timeout);
+URL_FILE	*url_fopen(const char *url, double timeout);
 
 int		url_fclose(URL_FILE *file),
 		url_feof(URL_FILE *file);
